requirements.txt
pdf2image==1.16.3
pytesseract==0.3.10
Pillow>=10.3.0
nltk==3.8.1
PyYAML==6.0.2
underthesea==6.8.0